requirements.txt
annotated-types==0.7.0
anyio==4.9.0
catboost==1.2.8
click==8.2.0
contourpy==1.3.2
cycler==0.12.1
fastapi==0.115.12
fonttools==4.58.0
graphviz==0.20.3
h11==0.16.0
idna==3.10
imbalanced-learn==0.13.0
joblib==1.5.0
kiwisolver==1.4.8
lightgbm==4.6.0
matplotlib==3.10.3
narwhals==1.39.0
nltk==3.9.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
sns==0.1
starlette==0.46.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
xgboost==3.0.0
